Fix a bug in gradient checker v2 for empty gradients where it led to a shape mismatch error.
The list comprehension "[x.dtype for x in xs]" overrides the value of x in the outer namespace (in python 2.x) and this leads to a wrong failure of the condition "grad.shape != x.shape".

Also added a unit test which would have failed without the fix.
Note: This bug arose while adding support for zero-sized inputs in EinsumOp.
PiperOrigin-RevId: 268054768