Lower TensorFlow Pack op to TensorFlow ExpandDims and ConcatV2 ops
Introduced a pass for testing but this kind of patterns lowering within TensorFlow are intended to be used with target specific patterns instead of in an isolated pass. That way ops that are not targeted directly can be lowered to simpler TensorFlow ops with conversion patterns for the target. Added a separate library to get TensorFlow lowering patterns.

PiperOrigin-RevId: 268804239