Add option to upgrade legacy control flow during import
Add a new option/flag to upgrade legacy TF behavior on import. This is to support some legacy models without needing to support the legacy behavior internally. This uses FunctionalizeControlFlow and is a new application of the pass (previously restricted to only being used when compiling for XLA), and it might not work for all cases.

Upgrading legacy graphs may also be split out into a separate tool finally.

PiperOrigin-RevId: 271046929