Unconditionally re-insert default attributes in GraphConstructor
GraphConstructor used to re-insert default attributes only when it
was emulating python's import_graph_def functionality. It should be
safe to re-insert always instead. This is needed so that whenever
we convert from GraphDef to Graph (a necessary step for execution)
default attributes are added.

The change also adds an option to perform NodeDef validation to
GraphConstructor. This option is needed when we want to convert
a GraphDef without default attributes to Graph with default attributes
and validate (e.g. in GraphMgr::InitItem). Validating as a separate step is:
 - more expensive
 - Existing validation routines work on GraphDefs, not Graphs. We can't
   use them because the GraphDef without default attrs is considered invalid.

This change is the first part of stripping and re-inserting default
attributes across WorkerService, which is needed to support forward
compatibility across RPCs, i.e. an old server trying to run a new graph.

PiperOrigin-RevId: 271168101