Replace Device::FillContextMap() with Device::TryGetContextMap().
Some devices provide a per-step DeviceContext to be used when executing kernels. At present, we use a "DeviceContextMap" that maps node IDs to DeviceContext* in the executor, but all existing devices return the same DeviceContext* for all nodes. This change switches the API to (optionally) return a single DeviceContext* that (if set) will be used by all nodes. As a result, we no longer need to store a Graph inside ExecutorImpl, which reduces the memory consumption for all executors.

PiperOrigin-RevId: 271877942